Randomly load an icon in all technical view nodes
The first intention was to do it with ::before and ::after selectors, but it seems that Cytoscape doesn't allow styling the nodes it generates that way.

Had to create an assets folder to add the three icons we were planning to use. Added via relative path.

This also fucks up the change to blue when selecting node, meaning it always stays grey, no matter what.